server/requirements.txt
# FastAPI and server dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Security dependencies
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-magic==0.4.27
slowapi==0.1.9
cryptography>=41.0.0

# YOLO and computer vision
ultralytics>=8.0.0
opencv-python>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0

# Detectron2 dependencies (optional - install separately if needed)
# detectron2 @ git+https://github.com/facebookresearch/detectron2.git
torch>=1.8.0
torchvision>=0.9.0

# Additional utilities
pydantic>=2.0.0
